requirements.txt
streamlit
openai
requests
plotly
beautifulsoup4
python-dotenv
